Small changes. Still function to count result does not work :/
