+++ frontend/src/App.js
@@ -2,7 +2,6 @@ import { BrowserRouter as Router, Routes, Route, useLocation  } from 'react-rout
 import FramesDistribution from './pages/FrameDistribution';
 import Comparison from './pages/Comparison';
 import Quiz from './pages/Quiz';
-import QuizNew from './pages/QuizNew';
 import NavigationTabs from './components/Navigation';
 import Menu from './pages/Menu';
 import { SettingsProvider } from './context/SettingsContext';
@@ -21,7 +20,7 @@ function Layout() {
                   <Route path="/" element={<Menu />} />
                   <Route path="/compress" element={<FramesDistribution />} />
                   <Route path="/comparison" element={<Comparison />} />
-                  <Route path="/quiz" element={<QuizNew />} />
+                  <Route path="/quiz" element={<Quiz />} />
               </Routes>
               </FramesProvider>
           </SettingsProvider>
